+++ playground/hmr/invalidation-circular-deps/invalidate-handled-in-circle/child.js
@@ -1,0 +1,9 @@
+import './parent'
+
+if (import.meta.hot) {
+  import.meta.hot.accept(() => {
+    import.meta.hot.invalidate()
+  })
+}
+
+export const value = 'child'