+++ playground/optimize-missing-deps/__test__/optimize-missing-deps.spec.ts
@@ -1,4 +1,3 @@
-import fetch from 'node-fetch'
 import { expect, test } from 'vitest'
 import { port } from './serve'
 import { isBuild, page, untilUpdated } from '~utils'